Add optional timeout (seconds) to cmd:Exec()
The hardcoded const Timeout value of 10 seconds is not useful for all but trivial operations. This patch adds an optional int64 value to the exec() signature so that the timeout may optionally be set to a non-zero value.

The default behavior has not been modified and is backwards API compatible.